🚚 api: remove controller to certification/session-management context
