+++ api/src/certification/session-management/application/session-controller.js
@@ -1,4 +1,7 @@
+import * as jurySessionSerializer from '../../../../lib/infrastructure/serializers/jsonapi/jury-session-serializer.js';
+import * as sessionValidator from '../../shared/domain/validators/session-validator.js';
 import { usecases } from '../domain/usecases/index.js';
+import * as jurySessionRepository from '../infrastructure/repositories/jury-session-repository.js';
 import * as sessionSerializer from '../infrastructure/serializers/session-serializer.js';
 
 const get = async function (request, h, dependencies = { sessionSerializer }) {
@@ -7,7 +10,27 @@ const get = async function (request, h, dependencies = { sessionSerializer }) {
   return dependencies.sessionSerializer.serialize({ session, hasSupervisorAccess, hasSomeCleaAcquired });
 };
 
+const findPaginatedFilteredJurySessions = async function (
+  request,
+  h,
+  dependencies = {
+    jurySessionRepository,
+    jurySessionSerializer,
+    sessionValidator,
+  },
+) {
+  const { filter, page } = request.query;
+  const normalizedFilters = dependencies.sessionValidator.validateAndNormalizeFilters(filter);
+  const jurySessionsForPaginatedList = await dependencies.jurySessionRepository.findPaginatedFiltered({
+    filters: normalizedFilters,
+    page,
+  });
+
+  return dependencies.jurySessionSerializer.serializeForPaginatedList(jurySessionsForPaginatedList);
+};
+
 const sessionController = {
+  findPaginatedFilteredJurySessions,
   get,
 };
